bsbd3-project/requirements.txt
Flask>=2.3.0
requests>=2.31.0
psycopg2-binary>=2.9.0
pandas>=2.0.0
cryptography>=41.0.0